+++ data-raw/extractfrommarinedatabase.R
@@ -1,9 +1,9 @@
 # R has to be within the working directory of this file
 
-# extracted from calib marine database on 2 April 2025:
+# extracted from calib marine database on 14 December 2025:
 # http://calib.org/marine/query select * from details
-# 1979 records
-marinedata <- read.csv("marinedatabase.csv") # 247 kB
+# 2054 records
+marinedata <- read.csv("marinedatabase.csv") # 290 kB
 
 # references: downloaded results of 'select * from refs' and saved as .csv file
 refs <- read.csv("refs.csv")